Conflicts:

	board/eukrea/cpu9260/cpu9260.c
	drivers/serial/serial_s5pc1xx.c
	include/asm-arm/arch-s5pc1xx/clock.h
	include/asm-arm/arch-s5pc1xx/gpio.h
	include/asm-arm/arch-s5pc1xx/pwm.h
	include/asm-arm/arch-s5pc1xx/uart.h
	include/configs/cpu9260.h
	include/configs/cpuat91.h
	include/configs/davinci_dm355evm.h
	include/linux/mtd/samsung_onenand.h

+++ include/asm-arm/arch-s5pc1xx/gpio.h
@@ -23,12 +23,12 @@
 
 #ifndef __ASSEMBLY__
 struct s5pc1xx_gpio_bank {
-	unsigned long	con;
-	unsigned long	dat;
-	unsigned long	pull;
-	unsigned long	drv;
-	unsigned long	pdn_con;
-	unsigned long	pdn_pull;
+	unsigned int	con;
+	unsigned int	dat;
+	unsigned int	pull;
+	unsigned int	drv;
+	unsigned int	pdn_con;
+	unsigned int	pdn_pull;
 	unsigned char	res1[8];
 };
